third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-constructor-expected.txt

Tests the OfflineAudioContext constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new OfflineAudioContext(1, 0, 44100) threw exception SyntaxError: Failed to construct 'OfflineAudioContext': number of frames cannot be zero..
PASS context = new OfflineAudioContext(2, 512, 2999) threw exception IndexSizeError: Failed to construct 'OfflineAudioContext': The sampleRate provided (2999) is outside the range [3000, 384000]..
PASS context = new OfflineAudioContext(2, 512, 3000) did not throw exception.
PASS context.length is 512
PASS context = new OfflineAudioContext(2, 1024, 384000) did not throw exception.
PASS context.length is 1024
PASS context = new OfflineAudioContext(2, 1024, 384001) threw exception IndexSizeError: Failed to construct 'OfflineAudioContext': The sampleRate provided (384001) is outside the range [3000, 384000]..
PASS context = new OfflineAudioContext(2, 2048, 8000) did not throw exception.
PASS context.length is 2048
PASS context = new OfflineAudioContext(2, 4096, 11025) did not throw exception.
PASS context.length is 4096
PASS context = new OfflineAudioContext(2, 512, 22050) did not throw exception.
PASS context.length is 512
PASS context = new OfflineAudioContext(2, 512, 44100) did not throw exception.
PASS context.length is 512
PASS context = new OfflineAudioContext(2, 512, 48000) did not throw exception.
PASS context.length is 512
PASS context = new OfflineAudioContext(2, 512, 88200) did not throw exception.
PASS context.length is 512
PASS context.length = 99 did not throw exception.
PASS context.length is 512
PASS context = new OfflineAudioContext(2, 512, 96000) did not throw exception.
PASS context = new OfflineAudioContext(1, 0, NaN) threw exception TypeError: Failed to construct 'OfflineAudioContext': The provided float value is non-finite..
PASS context = new OfflineAudioContext(1, 0, Infinity) threw exception TypeError: Failed to construct 'OfflineAudioContext': The provided float value is non-finite..
PASS context = new OfflineAudioContext(7, 512, 48000) did not throw exception.
PASS context.destination.channelCount is 7
PASS context.destination.maxChannelCount is 7
PASS context.destination.channelCountMode is "explicit"
PASS context.destination.channelInterpretation is "speakers"
PASS successfullyParsed is true

TEST COMPLETE
